Move filter_out_sequencer_commitments to primitives
